station/requirements.txt
fastapi[all]
psycopg2-binary
cryptography
pycryptodome
sqlalchemy
requests
jose
pyjwt
docker
numpy
pandas
python-dotenv
apache-airflow[postgres, celery]
minio
pytorch-lightning
pytest
torchvision
python-keycloak
git+https://gitlab.com/PersonalHealthTrain/implementations/germanmii/difuture/train-container-library.git
git+https://gitlab.com/PersonalHealthTrain/implementations/germanmii/difuture/incubating/pht/conductor-library.git